+++ app/javascript/App/components/NewsfeedPage/NewsBody.tsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import * as React from 'react'
 import timeago from 'timeago.js'
 import sanitizeHtml from 'sanitize-html'
 import _ from 'lodash'
@@ -6,49 +6,61 @@ import CoinTags from '../CoinTags'
 import BulletSpacer from '../BulletSpacer'
 import Icon from '../Icon'
 
-export default class NewsBody extends Component {
+import { NewsItem } from './types';
+
+interface Category {
+  name: string,
+};
+
+interface NewsBodyProps {
+  newsItem: NewsItem,
+  categories: Array<Category>,
+};
+
+export default class NewsBody extends React.Component<NewsBodyProps, {}>  {
   render() {
-    const {
-      selectNewsItemFromList,
-      activeEntity,
-      selectNewsCategories,
-    } = this.props
-    const { id } = activeEntity
-    const newsItem = selectNewsItemFromList(id)
+    // const {
+    //   selectNewsItemFromList,
+    //   activeEntity,
+    //   selectNewsCategories,
+    // } = this.props
+
+    const { newsItem, categories } = this.props;
+  
     if (!newsItem) {
       return null
     }
-    const categories = selectNewsCategories(newsItem)
-    const content =
-      _.trim(newsItem.get('content')) || _.trim(newsItem.get('summary'))
+
+    const content = _.trim(newsItem.content) || _.trim(newsItem.summary);
+
     return (
       <div className="pa3 bg-white min-h-100 selected-news-content">
         <CoinTags itemWithCoinLinkData={newsItem} />
-        <h1 className="break-word f4">{newsItem.get('title')}</h1>
+        <h1 className="break-word f4">{newsItem.title}</h1>
         <div className="mb3 f6">
           <a
-            href={newsItem.get('url')}
+            href={newsItem.url}
             target="_blank"
             rel="nofollow"
             className="break-all"
           >
             <Icon name="link" className="mr1 f7" regular />
-            {newsItem.get('url')}
+            {newsItem.url}
           </a>
         </div>
         <div className="mb3 f6">
           <Icon name="clock" className="mr1 f7" regular />
-          {timeago().format(newsItem.get('feed_item_published_at'))}
+          {timeago().format(newsItem.feed_item_published_at)}
           <BulletSpacer />
           <span>
-            {new Date(newsItem.get('feed_item_published_at')).toLocaleString()}
+            {new Date(newsItem.feed_item_published_at).toLocaleString()}
           </span>
         </div>
-        {categories.size > 0 && (
+        {categories.length > 0 && (
           <div className="mv3">
             {categories.map((category, index) => (
               <div key={index} className="tag-alt">
-                {category.get('name')}
+                {category.name}
               </div>
             ))}
           </div>
