+++ app/javascript/App/components/NewsfeedPage/types.ts
@@ -1,0 +1,7 @@
+export interface NewsItem {
+  title: string
+  summary: string
+  content: string
+  url: string
+  feed_item_published_at: string
+}
